Extract Check/report formatting helpers into taskpane-check-formatters.js
Move five pure formatting helpers out of taskpane.js into a new
dedicated module. No calculation or Excel write logic is touched.

Helpers moved:
  - formatSkippedCandidateDetail
  - checkMetricTypesFromBlocks
  - formatCheckUserVisibleIssues
  - formatCheckCalculationBlocks
  - formatCheckBannerSummary
